fix(s3api): cast math.MaxInt64 to int64 for 32-bit vet
On 32-bit platforms, math.MaxInt64 is an untyped int constant that
overflows int (32-bit) when used directly in fmt.Errorf with %d.
Cast to int64 explicitly to fix Go Vet 32-bit.

+++ weed/s3api/s3api_bucket_quota_handlers.go
@@ -230,7 +230,7 @@ func convertQuotaToBytes(size int64, unit string) (int64, error) {
 		return 0, fmt.Errorf("unsupported quota_unit %q", unit)
 	}
 	if multiplier > 0 && size > math.MaxInt64/multiplier {
-		return 0, fmt.Errorf("quota_size %d %s overflows maximum bytes (%d)", size, unit, math.MaxInt64)
+		return 0, fmt.Errorf("quota_size %d %s overflows maximum bytes (%d)", size, unit, int64(math.MaxInt64))
 	}
 	return size * multiplier, nil
 }
